# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для отправки напоминаний на email",
        "title": "TTL Email Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@ttl-checker.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/email-scheduler/reminder": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email-scheduler"
                ],
                "summary": "Просмотр напоминаний",
                "operationId": "get-reminders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статус напоминания",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/email-scheduler/reminder/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email-scheduler"
                ],
                "summary": "Создание нового напоминания",
                "operationId": "reminder-create",
                "parameters": [
                    {
                        "description": "Информация для напоминания",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Reminder"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/email-scheduler/reminder/create-range": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email-scheduler"
                ],
                "summary": "Создание временного диапазона для напоминаний",
                "operationId": "reminder-create-range",
                "parameters": [
                    {
                        "description": "Информация для создания диапазона напоминаний",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RemindersWithTimeRange"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/email-scheduler/reminder/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email-scheduler"
                ],
                "summary": "Удалить напоминание",
                "operationId": "delete-reminder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id reminder",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "Проверка прав пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация пользователя",
                "operationId": "sign-in",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserSignInRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Создание нового пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "sign-up",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/verify-email/send": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Отправка подтверждения почты пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Отправка подтверждения почты пользователя",
                "operationId": "token-send",
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Reminder": {
            "type": "object",
            "required": [
                "send_at",
                "subject",
                "text"
            ],
            "properties": {
                "send_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "default": "pending"
                },
                "subject": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.RemindersWithTimeRange": {
            "type": "object",
            "required": [
                "range_end",
                "range_start",
                "repeat_condition",
                "subject",
                "text"
            ],
            "properties": {
                "condition": {
                    "type": "string"
                },
                "range_end": {
                    "type": "string"
                },
                "range_start": {
                    "type": "string"
                },
                "repeat_condition": {
                    "type": "string",
                    "enum": [
                        "year",
                        "month",
                        "week",
                        "day"
                    ]
                },
                "subject": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "registered_date": {
                    "type": "string"
                }
            }
        },
        "models.UserSignInRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "default": "marisasha228@bk.ru"
                },
                "password": {
                    "type": "string",
                    "default": "123"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Введите: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}